feat: Refactor reset password and update user data handling
This commit refactors the `resetPassword` functionality, moving
the logic to the service layer and improving the response data.
It also updates user selection and omits sensitive fields based
on the user role after login and password reset. Additionally,
the `looking_for` field in the user schema was renamed to
`genre` and `TModelZod` was updated accordingly.